feat(login): Add JWT auto-renew failure callback.
Sometimes we can't renew our JWT before it expires One possible scenario where
this could happen is if the user's computer is suspended when the JWT expires,
upon coming back from suspension we try to renew the JWT using a now-invalid
JWT.  This leaves us in a limbo state where we have no valid JWT and no way
to retrieve a new one.

Add a callback in case auto-renewing a JWT fails, so that we can at least
inform the user that the application will no longer work as expected.

Supports EN-5289.

Signed-off-by: Jeff Cuevas-Koch <jcuevas-koch@gmvsync.com>