feat(voip): Expose isMuted on VOIP heartbeats.
VOIP heartbeats sent to users contain a `desiredCallState` that informs
the user whether they should connect to a call.  We've added an
`isMuted` property to these heartbeats to support listen-only call
connections.

This commit exposes the `isMuted` property to VOIP heartbeat handlers.

I also updated the mocks that refer to Users and Vehicles on call
participants to accurately match the Track API models.

Contributes to EN-7644.

Signed-off-by: Jeff Cuevas-Koch <jcuevas-koch@gmvsync.com>